refactor tracer, move copy-paste to dedicated file
