+++ setup.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.2.0
+current_version = 0.1.1
 commit = True
 tag = True
 
@@ -7,7 +7,7 @@ tag = True
 search = version='{current_version}'
 replace = version='{new_version}'
 
-[bumpversion:file:celery_executor/__init__.py]
+[bumpversion:file:celery_pool_asyncio/__init__.py]
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
 
@@ -22,11 +22,11 @@ test = pytest
 
 [tool:pytest]
 collect_ignore = ['setup.py']
-python_files = tests.py test_*.py tests/*.py celery_executor/*
+python_files = tests.py test_*.py tests/*.py celery_pool_asyncio/*
 norecursedirs = .git conf src .eggs
 doctest_optionflags = NORMALIZE_WHITESPACE ALLOW_UNICODE ALLOW_BYTES
 addopts = 
 	-v -s
-	--doctest-modules
-	--cov --cov-report=term-missing --cov-report=html --cov-config=.coveragerc
+#	--doctest-modules
+#	--cov --cov-report=term-missing --cov-report=html --cov-config=.coveragerc
